# LctMonolith/openapi-gamification.yaml
openapi: 3.0.3
info:
  title: LctMonolith Gamification API
  version: 1.0.0
  description: |
    Comprehensive REST API for gamification module (players, missions, ranks, skills, rewards, dialogue, inventory, store, analytics, notifications, auth).
    Authentication via JWT Bearer (Authorization: Bearer <token>). Admin endpoints require role=Admin.
servers:
  - url: https://localhost:5001
    description: Local HTTPS
  - url: http://localhost:5000
    description: Local HTTP
security:
  - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Id: { type: string, format: uuid }
...
    Rank:
      type: object
      properties:
        id: { $ref: '#/components/schemas/Id' }
        title: { type: string }
        expNeeded: { type: integer }
...
    PlayerProgress:
      type: object
      properties:
        playerId: { $ref: '#/components/schemas/Id' }
        playerName: { type: string }
        currentRank: { $ref: '#/components/schemas/Rank', nullable: true }
        totalExperience: { type: integer }
        totalMana: { type: integer }
        completedMissions: { type: integer }
        totalAvailableMissions: { type: integer }
        skillLevels:
          type: object
          additionalProperties: { type: integer }
...
    Profile:
      type: object
      properties:
        id: { $ref: '#/components/schemas/Id' }
        userId: { $ref: '#/components/schemas/Id' }
        firstName: { type: string, nullable: true }
        lastName: { type: string, nullable: true }
        birthDate: { type: string, format: date, nullable: true }
        about: { type: string, nullable: true }
        location: { type: string, nullable: true }
        avatarUrl: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
paths:
...
  /api/players/{playerId}:
    get:
      tags: [Players]
      security: [ { bearerAuth: [] } ]
      summary: Get player by id
      parameters:
        - in: path
          name: playerId
          required: true
          schema: { $ref: '#/components/schemas/Id' }
      responses:
        '200': { description: OK }
        '404': { description: Not Found }
  /api/players/{playerId}/progress:
    get:
      tags: [Players]
      security: [ { bearerAuth: [] } ]
      summary: Get player overall progress
      parameters:
        - in: path
          name: playerId
          required: true
          schema: { $ref: '#/components/schemas/Id' }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/PlayerProgress' } } } }
        '404': { description: Not Found }
  /api/players/user/{userId}:
    get:
      tags: [Players]
      security: [ { bearerAuth: [] } ]
      summary: Get player by user id
      parameters:
        - in: path
          name: userId
          required: true
          schema: { $ref: '#/components/schemas/Id' }
      responses:
        '200': { description: OK }
        '404': { description: Not Found }
  /api/players:
    post:
      tags: [Players]
      security: [ { bearerAuth: [] } ]
      summary: Create player (Admin)
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                userId: { $ref: '#/components/schemas/Id' }
                username: { type: string }
              required: [userId,username]
      responses:
        '201': { description: Created }
      x-roles: [Admin]
  /api/players/{playerId}/experience:
    post:
      tags: [Players]
      security: [ { bearerAuth: [] } ]
      summary: Adjust player experience (Admin)
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                value: { type: integer }
              required: [value]
      parameters:
        - in: path
          name: playerId
          required: true
          schema: { $ref: '#/components/schemas/Id' }
      responses:
        '200': { description: OK }
      x-roles: [Admin]
  /api/players/{playerId}/mana:
    post:
      tags: [Players]
      security: [ { bearerAuth: [] } ]
      summary: Adjust player mana (Admin)
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                value: { type: integer }
              required: [value]
      parameters:
        - in: path
          name: playerId
          required: true
          schema: { $ref: '#/components/schemas/Id' }
      responses:
        '200': { description: OK }
      x-roles: [Admin]
  /api/players/top:
    get:
      tags: [Players]
      security: [ { bearerAuth: [] } ]
      summary: Top players by experience
      parameters:
        - in: query
          name: count
          schema: { type: integer, default: 10 }
      responses:
        '200': { description: OK }
  /api/ranks:
    get:
      tags: [Ranks]
      security: [ { bearerAuth: [] } ]
      summary: List ranks
      responses: { '200': { description: OK } }
    post:
      tags: [Ranks]
      security: [ { bearerAuth: [] } ]
      summary: Create rank (Admin)
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title: { type: string }
                expNeeded: { type: integer }
              required: [title,expNeeded]
      responses: { '201': { description: Created } }
      x-roles: [Admin]
  /api/ranks/{id}:
    get:
      tags: [Ranks]
      security: [ { bearerAuth: [] } ]
      summary: Get rank
      parameters:
        - in: path
          name: id
          required: true
          schema: { $ref: '#/components/schemas/Id' }
      responses: { '200': { description: OK }, '404': { description: Not Found } }
    put:
      tags: [Ranks]
      security: [ { bearerAuth: [] } ]
      summary: Update rank (Admin)
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title: { type: string }
                expNeeded: { type: integer }
              required: [title,expNeeded]
      responses: { '200': { description: OK }, '404': { description: Not Found } }
      x-roles: [Admin]
    delete:
      tags: [Ranks]
      security: [ { bearerAuth: [] } ]
      summary: Delete rank (Admin)
      responses: { '204': { description: No Content }, '404': { description: Not Found } }
      x-roles: [Admin]
  /api/ranks/validate-advance/{playerId}/{targetRankId}:
    get:
      tags: [Ranks]
      security: [ { bearerAuth: [] } ]
      summary: Validate advancement
      parameters:
        - in: path
          name: playerId
          required: true
          schema: { $ref: '#/components/schemas/Id' }
        - in: path
          name: targetRankId
          required: true
          schema: { $ref: '#/components/schemas/Id' }
      responses: { '200': { description: OK } }
...
  /api/skills/player/{playerId}:
    get:
      tags: [Skills]
      security: [ { bearerAuth: [] } ]
      summary: List player skills
      parameters:
        - in: path
          name: playerId
          required: true
          schema: { $ref: '#/components/schemas/Id' }
      responses: { '200': { description: OK } }
  /api/skills/player/{playerId}/{skillId}:
    post:
      tags: [Skills]
      security: [ { bearerAuth: [] } ]
      summary: Update player skill (Admin)
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                level: { type: integer }
              required: [level]
      parameters:
        - in: path
          name: playerId
          required: true
          schema: { $ref: '#/components/schemas/Id' }
        - in: path
          name: skillId
          required: true
          schema: { $ref: '#/components/schemas/Id' }
      responses: { '200': { description: OK } }
      x-roles: [Admin]
...
  /api/rewards/mission/{missionId}/skills:
    get:
      tags: [Rewards]
      security: [ { bearerAuth: [] } ]
      summary: Mission skill rewards
      parameters:
        - in: path
          name: missionId
          required: true
          schema: { $ref: '#/components/schemas/Id' }
      responses: { '200': { description: OK } }
  /api/rewards/mission/{missionId}/items:
    get:
      tags: [Rewards]
      security: [ { bearerAuth: [] } ]
      summary: Mission item rewards
      parameters:
        - in: path
          name: missionId
          required: true
          schema: { $ref: '#/components/schemas/Id' }
      responses: { '200': { description: OK } }
  /api/rewards/mission/{missionId}/can-claim/{playerId}:
    get:
      tags: [Rewards]
      security: [ { bearerAuth: [] } ]
      summary: Can claim mission rewards
      parameters:
        - in: path
          name: missionId
          required: true
          schema: { $ref: '#/components/schemas/Id' }
        - in: path
          name: playerId
          required: true
          schema: { $ref: '#/components/schemas/Id' }
      responses: { '200': { description: OK } }
  /api/rewards/mission/{missionId}/claim:
    post:
      tags: [Rewards]
      security: [ { bearerAuth: [] } ]
      summary: Claim mission rewards
      parameters:
        - in: path
          name: missionId
          required: true
          schema: { $ref: '#/components/schemas/Id' }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                playerId: { $ref: '#/components/schemas/Id' }
              required: [playerId]
      responses: { '200': { description: OK }, '409': { description: Conflict } }
  /api/rewards/mission/{missionId}/force-distribute:
    post:
      tags: [Rewards]
      security: [ { bearerAuth: [] } ]
      summary: Force distribute mission rewards (Admin)
      parameters:
        - in: path
          name: missionId
          required: true
          schema: { $ref: '#/components/schemas/Id' }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                playerId: { $ref: '#/components/schemas/Id' }
              required: [playerId]
      responses: { '200': { description: OK } }
      x-roles: [Admin]
...
  /api/profile/me:
    get:
      tags: [Profile]
      security: [ { bearerAuth: [] } ]
      summary: Get current user profile
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/Profile' } } } }
        '404': { description: Not Found }
  /api/profile/{userId}:
    get:
      tags: [Profile]
      security: [ { bearerAuth: [] } ]
      summary: Get profile by user id (Admin)
      parameters:
        - in: path
          name: userId
          required: true
          schema: { $ref: '#/components/schemas/Id' }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/Profile' } } } }
        '404': { description: Not Found }
      x-roles: [Admin]
  /api/profile:
    put:
      tags: [Profile]
      security: [ { bearerAuth: [] } ]
      summary: Upsert current user profile
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                firstName: { type: string, nullable: true }
                lastName: { type: string, nullable: true }
                birthDate: { type: string, format: date, nullable: true }
                about: { type: string, nullable: true }
                location: { type: string, nullable: true }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/Profile' } } } }
  /api/profile/avatar:
    post:
      tags: [Profile]
      security: [ { bearerAuth: [] } ]
      summary: Upload avatar image (multipart/form-data)
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200': { description: OK, content: { application/json: { schema: { type: object, properties: { avatarUrl: { type: string } } } } } }
        '400': { description: Bad Request }
    delete:
      tags: [Profile]
      security: [ { bearerAuth: [] } ]
      summary: Delete avatar image
      responses:
        '204': { description: No Content }
        '404': { description: Not Found }